refactor: update distance and duration in backend

+++ src/main/java/at/technikum_wien/tourplannerapi/dto/TourDTO.java
@@ -15,7 +15,7 @@ public class TourDTO {
     private String toLocation;
     private String transportType;
     private double distance;
-    private double estimatedTime;
+    private long estimatedTime;
     private int popularity;
     private double childFriendliness;
